+++ android/build.gradle
@@ -60,12 +60,25 @@ dependencies {
     implementation "androidx.core:core-ktx:1.13.1"
     implementation "androidx.localbroadcastmanager:localbroadcastmanager:1.1.0"
     implementation "androidx.lifecycle:lifecycle-process:2.5.1"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3"
     implementation 'androidx.media:media:1.7.0'
 
+    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.7.0")
+    implementation("androidx.activity:activity-compose:1.9.2")
+    implementation("androidx.compose.ui:ui")
+    implementation("androidx.compose.ui:ui-graphics")
+    implementation("androidx.compose.ui:ui-tooling-preview")
+    implementation("androidx.compose.material3:material3")
+    implementation("androidx.media3:media3-exoplayer:1.4.1")
+    implementation("androidx.media3:media3-session:1.4.1")
+    implementation("androidx.media3:media3-ui:1.4.1")
+    implementation("androidx.media3:media3-exoplayer-hls:1.4.1")
+    implementation("androidx.media3:media3-exoplayer-dash:1.4.1")
+    implementation("androidx.media3:media3-exoplayer-smoothstreaming:1.4.1")
+    implementation("androidx.media3:media3-common:1.4.1")
+
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3"
+
     implementation 'io.coil-kt:coil:2.2.0'
-    api 'com.google.android.exoplayer:exoplayer:2.19.1'
-    api 'com.google.android.exoplayer:extension-mediasession:2.19.1'
     api 'com.jakewharton.timber:timber:5.0.1'
 
     implementation 'androidx.test:rules:1.5.0'
