--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/AAMediaSessionCallBack.kt
@@ -1,10 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-import android.os.Bundle
-
-interface AAMediaSessionCallBack {
-    fun handlePlayFromMediaId(mediaId: String?, extras: Bundle?)
-    fun handlePlayFromSearch(query: String?, extras: Bundle?)
-    fun handleSkipToQueueItem(id: Long)
-    fun handleCustomActions(action: String?, extras: Bundle?)
-}
