--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/AudioContentType.kt
@@ -1,10 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-enum class AudioContentType {
-    MUSIC,
-    SPEECH,
-    SONIFICATION,
-    MOVIE,
-    UNKNOWN
-}
-
