--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/BufferConfig.kt
@@ -1,8 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-data class BufferConfig(
-    val minBuffer: Int?,
-    val maxBuffer: Int?,
-    val playBuffer: Int?,
-    val backBuffer: Int?,
-)
