--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/CacheConfig.kt
@@ -1,17 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-/**
- * Configuration for cache properties of player.
- */
-data class CacheConfig(
-    /**
-     * Maximum player cache size in kilobytes.
-     */
-    val maxCacheSize: Long?,
-
-    /**
-     * Cache identifier, used to make cache directory.
-     */
-    val identifier: String = "TrackPlayer"
-)
-
