--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/Capability.kt
@@ -1,19 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-enum class Capability {
-    PLAY,
-    PLAY_FROM_ID,
-    PLAY_FROM_SEARCH,
-    PAUSE,
-    STOP,
-    SEEK_TO,
-    SKIP,
-    SKIP_TO_NEXT,
-    SKIP_TO_PREVIOUS,
-    JUMP_FORWARD,
-    JUMP_BACKWARD,
-    SET_RATING,
-    LIKE,
-    DISLIKE,
-    BOOKMARK
-}
