--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/FocusChangeData.kt
@@ -1,3 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-data class FocusChangeData(val isPaused: Boolean, val isFocusLostPermanently: Boolean)
