--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/PlayWhenReadyChangeData.kt
@@ -1,5 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-import com.google.android.exoplayer2.Player
-
-data class PlayWhenReadyChangeData(val playWhenReady: Boolean, val pausedBecauseReachedEnd: Boolean)
