--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/PlaybackError.kt
@@ -1,6 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-data class PlaybackError (
-    val code: String? = null,
-    val message: String? = null
-)
