--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/PlayerOptions.kt
@@ -1,9 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-interface PlayerOptions {
-    var alwaysPauseOnInterruption: Boolean
-}
-
-internal data class DefaultPlayerOptions(
-    override var alwaysPauseOnInterruption: Boolean = false,
-) : PlayerOptions
