--- android/src/main/java/com/doublesymmetry/kotlinaudio/models/WakeMode.kt
@@ -1,7 +1,0 @@
-package com.doublesymmetry.kotlinaudio.models
-
-enum class WakeMode {
-    NONE,
-    LOCAL,
-    NETWORK,
-}
