--- android/src/main/java/com/doublesymmetry/kotlinaudio/players/AudioPlayer.kt
@@ -1,9 +1,0 @@
-package com.doublesymmetry.kotlinaudio.players
-
-import android.content.Context
-import com.doublesymmetry.kotlinaudio.models.BufferConfig
-import com.doublesymmetry.kotlinaudio.models.CacheConfig
-import com.doublesymmetry.kotlinaudio.models.PlayerConfig
-import com.doublesymmetry.kotlinaudio.models.AAMediaSessionCallBack
-
-class AudioPlayer(context: Context, playerConfig: PlayerConfig = PlayerConfig(), bufferConfig: BufferConfig? = null, cacheConfig: CacheConfig? = null, mediaSessionCallback: AAMediaSessionCallBack): BaseAudioPlayer(context, playerConfig, bufferConfig, cacheConfig, mediaSessionCallback)
