+++ android/src/main/java/com/lovegaoshi/kotlinaudio/event/EventHolder.kt
@@ -1,0 +1,30 @@
+package com.lovegaoshi.kotlinaudio.event
+
+class EventHolder internal constructor(private val playerEventHolder: PlayerEventHolder) {
+    val audioItemTransition
+        get() = playerEventHolder.audioItemTransition
+
+    val onAudioFocusChanged
+        get() = playerEventHolder.onAudioFocusChanged
+
+    val onCommonMetadata
+        get() = playerEventHolder.onCommonMetadata
+
+    val onTimedMetadata
+        get() = playerEventHolder.onTimedMetadata
+
+    val onPlayerActionTriggeredExternally
+        get() = playerEventHolder.onPlayerActionTriggeredExternally
+
+    val playbackEnd
+        get() = playerEventHolder.playbackEnd
+
+    val playWhenReadyChange
+        get() = playerEventHolder.playWhenReadyChange
+
+    val stateChange
+        get() = playerEventHolder.stateChange
+
+    val playbackError
+        get() = playerEventHolder.playbackError
+}
