+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/AudioContentType.kt
@@ -1,0 +1,10 @@
+package com.lovegaoshi.kotlinaudio.models
+
+enum class AudioContentType {
+    MUSIC,
+    SPEECH,
+    SONIFICATION,
+    MOVIE,
+    UNKNOWN
+}
+
