+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/BufferConfig.kt
@@ -1,0 +1,8 @@
+package com.lovegaoshi.kotlinaudio.models
+
+data class BufferConfig(
+    val minBuffer: Int?,
+    val maxBuffer: Int?,
+    val playBuffer: Int?,
+    val backBuffer: Int?,
+)
