+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/CacheConfig.kt
@@ -1,0 +1,17 @@
+package com.lovegaoshi.kotlinaudio.models
+
+/**
+ * Configuration for cache properties of player.
+ */
+data class CacheConfig(
+    /**
+     * Maximum player cache size in kilobytes.
+     */
+    val maxCacheSize: Long?,
+
+    /**
+     * Cache identifier, used to make cache directory.
+     */
+    val identifier: String = "TrackPlayer"
+)
+
