+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/Capability.kt
@@ -1,0 +1,19 @@
+package com.lovegaoshi.kotlinaudio.models
+
+enum class Capability {
+    PLAY,
+    PLAY_FROM_ID,
+    PLAY_FROM_SEARCH,
+    PAUSE,
+    STOP,
+    SEEK_TO,
+    SKIP,
+    SKIP_TO_NEXT,
+    SKIP_TO_PREVIOUS,
+    JUMP_FORWARD,
+    JUMP_BACKWARD,
+    SET_RATING,
+    LIKE,
+    DISLIKE,
+    BOOKMARK
+}
