+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/FocusChangeData.kt
@@ -1,0 +1,3 @@
+package com.lovegaoshi.kotlinaudio.models
+
+data class FocusChangeData(val isPaused: Boolean, val isFocusLostPermanently: Boolean)
