+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/PlayWhenReadyChangeData.kt
@@ -1,0 +1,3 @@
+package com.lovegaoshi.kotlinaudio.models
+
+data class PlayWhenReadyChangeData(val playWhenReady: Boolean, val pausedBecauseReachedEnd: Boolean)
