+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/PlaybackEndedReason.kt
@@ -1,0 +1,5 @@
+package com.lovegaoshi.kotlinaudio.models
+
+enum class PlaybackEndedReason {
+    PLAYED_UNTIL_END, PLAYER_STOPPED, SKIPPED_TO_NEXT, SKIPPED_TO_PREVIOUS, JUMPED_TO_INDEX
+}
