+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/PlaybackError.kt
@@ -1,0 +1,6 @@
+package com.lovegaoshi.kotlinaudio.models
+
+data class PlaybackError (
+    val code: String? = null,
+    val message: String? = null
+)
