+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/WakeMode.kt
@@ -1,0 +1,7 @@
+package com.lovegaoshi.kotlinaudio.models
+
+enum class WakeMode {
+    NONE,
+    LOCAL,
+    NETWORK,
+}
